+++ scenarios/game_of_tag_demo_map/scenario.py
@@ -29,24 +29,24 @@ seed(42)
 
 # training missions
 ego_missions = [
-    # t.EndlessMission(
-    #     begin=("top_left", 0, 15) # pred 1
-    # ),
+    t.EndlessMission(
+        begin=("top_left", 0, 15) # pred 1
+    ),
     t.EndlessMission(
         begin=("top", 1, 5) # pred 2
     ),
-    # t.EndlessMission(
-    #     begin=("top_left", 1, 15) # pred 3
-    # ),
-    # t.EndlessMission(
-    #     begin=("top", 3, 30) # pred 4
-    # ),
+    t.EndlessMission(
+        begin=("top_left", 1, 15) # pred 3
+    ),
+    t.EndlessMission(
+        begin=("top", 3, 30) # pred 4
+    ),
     t.EndlessMission(
         begin=("top", 2, 30) # prey 1
     ),
-    # t.EndlessMission(
-    #     begin=("top", 0, 5) # prey 2
-    # ),   
+    t.EndlessMission(
+        begin=("top", 0, 5) # prey 2
+    ),   
 ]
